Browser Crawler split
Split up the crawler and scraper to create more specialized module's.
